Fix broken compile This removes the BOM (Byte-Order Mark) from MainActivity.java. I'm pretty sure that this fixes the infamous "Cannot compile when MainActivity.java is open error" Alright, explanation time. A long, long, time ago, in a galaxy far far away (sorry, I just saw Star Wars Episode VII the other day), I edited the file using Notepad++ on my computer, and changed the encoding for some reason. When I did this, I set it to "UTF-8" encoding, which added a BOM which is unnecessary and unsupported by the Android toolchain.
This resulted that AIDE, the IDE I was using at the time, would see the BOM and report "ERROR: Unexpected", which would point to just before the "beginning" of the file, where this character was found.
It is most likely that AIDE does something interesting when the file isn't open in the IDE that allowed it to compile without error.
Why it worked in Eclipse, I have no idea.

I found this issue when II tried to compile under Android Studio and it reported “illegal character: \65279”. For future reference, this means that there is a BOM where there should not be one.

If this does remove that bug, I am so very glad. However, I must thank that bug, for it made a great story I could tell, and because it was the issue that finally pushed me to use Eclipse, which eventually lead to Android Studio, both of which are far more appropriate and much more productive than AIDE.
Therefore, to my infamous "'ERROR: Unexpected' when MainActivity.java is open for editing" bug, assuming this fixes you, I bid thee a hearty, melancholy, and bittersweet goodbye.

So long, old friend, I hope not to have to run into you again!